refactor : separate note form as component
